+++ Plugins/CrashRecovery/CHANGELOG.md
@@ -1,4 +1,7 @@
 # [CrashRecovery](https://1lighty.github.io/BetterDiscordStuff/?plugin=CrashRecovery "CrashRecovery") Changelog
+### 0.1.7
+- Changed to module.exports because useless backwards incompatbile changes are the motto for BBD apparently.
+
 ### 0.1.6
 - Removed usage of soon to be deprecated globals.